requirements.txt
python-telegram-bot==21.1
httpx==0.27.0
